% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeUnivariate.R
\name{imputeUnivariate}
\alias{imputeUnivariate}
\title{Univariate Imputation}
\usage{
imputeUnivariate(x)
}
\arguments{
\item{x}{A vector of any type possibly containing missing values.}
}
\value{
A vector of the same length and type as \code{x} but without missing values.
}
\description{
Fills missing values of a vector of any type by sampling with replacement from the non-missing values. Requires at least one non-missing value to run.
}
\examples{
imputeUnivariate(c(NA, 0, 1, 0, 1))
imputeUnivariate(c("A", "A", NA))
}
