% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.missRanger}
\alias{print.missRanger}
\title{Print Method}
\usage{
\method{print}{missRanger}(x, ...)
}
\arguments{
\item{x}{An object of class "missRanger".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for an object of class "missRanger".
}
\examples{
CO2_ <- generateNA(CO2, seed = 1)
imp <- missRanger(CO2_, data_only = FALSE)
imp
}
