% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missRanger.R
\name{revert}
\alias{revert}
\title{Revert conversion.}
\usage{
revert(con, X = con$X)
}
\arguments{
\item{con}{A list returned by \code{convert()}.}

\item{X}{A data frame with some columns to be converted back according to the
information stored in \code{converted}.}
}
\value{
A data frame.
}
\description{
Reverts conversions done by \code{convert()}.
}
