% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.missRanger}
\alias{summary.missRanger}
\title{Summary Method}
\usage{
\method{summary}{missRanger}(object, ...)
}
\arguments{
\item{object}{An object of class "missRanger".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Summary method for an object of class "missRanger".
}
\examples{
CO2_ <- generateNA(CO2, seed = 1)
imp <- missRanger(CO2_, pmm.k = 5, data_only = FALSE, num.threads = 1)
summary(imp)
}
