% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missRanger.R
\name{typeof2}
\alias{typeof2}
\title{A version of \code{typeof} internally used by \code{missRanger}.}
\usage{
typeof2(object)
}
\arguments{
\item{object}{Any object.}
}
\value{
A string.
}
\description{
Returns either "numeric" (double or integer), "factor", "character", "logical", "special" (mode numeric, but neither double nor integer) or "" (otherwise).
\code{missRanger} requires this information to deal with response types not natively supported by \code{ranger}.
}
\author{
Michael Mayer
}
