% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allVarsTwoSided.R
\name{allVarsTwoSided}
\alias{allVarsTwoSided}
\title{Extraction of Variable Names from Two-Sided Formula.}
\usage{
allVarsTwoSided(formula, data)
}
\arguments{
\item{formula}{A two-sided formula object.}

\item{data}{A \code{data.frame}.}
}
\value{
A \code{list} with two character vectors of variable names.
}
\description{
Takes a formula and a data frame and returns all variable names in both the lhs and the rhs. 
lrs and rhs are evaluated separately.
}
\examples{
allVarsTwoSided(Species + Sepal.Width ~ Petal.Width, iris)
allVarsTwoSided(. ~ ., iris)
allVarsTwoSided(.-Species ~ Sepal.Width, iris)
allVarsTwoSided(. ~ Sepal.Width, iris)
}
