% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNA.R
\name{generateNA}
\alias{generateNA}
\title{Adds Missing Values to a Data Set}
\usage{
generateNA(data, p = 0.1, seed = NULL)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{p}{Proportion of missing values to approximatly add to each column of \code{data}.}

\item{seed}{An integer seed.}
}
\value{
\code{data} with missing values.
}
\description{
Takes a data frame and replaces randomly part of the values by missing values.
}
\examples{
head(generateNA(iris))
}
